@@ -3,24 +3,24 @@
     <modelVersion>4.0.0</modelVersion>
     <groupId>gov.nist.itl.ssd.wipp</groupId>
     <artifactId>wipp-backend-argo-workflows</artifactId>
-    <version>3.1.0-SNAPSHOT</version>
+    <version>3.2.0-SNAPSHOT</version>
 
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <maven.compiler.source>1.8</maven.compiler.source>
-        <maven.compiler.target>1.8</maven.compiler.target>
+        <maven.compiler.release>21</maven.compiler.release>
+        <java.version>21</java.version>
     </properties>
 
     <dependencies>
         <dependency>
             <groupId>com.fasterxml.jackson.dataformat</groupId>
             <artifactId>jackson-dataformat-yaml</artifactId>
-            <version>2.10.0</version>
+            <version>2.17.0</version>
         </dependency>
         <dependency>
             <groupId>com.fasterxml.jackson.core</groupId>
             <artifactId>jackson-databind</artifactId>
-            <version>2.12.7.1</version>
+            <version>2.17.0</version>
         </dependency>
         <dependency>
             <groupId>gov.nist.itl.ssd.wipp</groupId>
@@ -28,4 +28,5 @@
             <version>${project.version}</version>
         </dependency>
     </dependencies>
+
 </project>
