@@ -4,15 +4,18 @@ import gov.nist.itl.ssd.wipp.backend.Application;
 import gov.nist.itl.ssd.wipp.backend.app.SecurityConfig;
 import gov.nist.itl.ssd.wipp.backend.core.model.computation.Plugin;
 import gov.nist.itl.ssd.wipp.backend.core.model.computation.PluginRepository;
-import org.junit.Before;
-import org.junit.Test;
-import org.junit.runner.RunWith;
+
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.TestInstance;
+import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.hateoas.MediaTypes;
 import org.springframework.security.test.context.support.WithAnonymousUser;
 import org.springframework.security.test.context.support.WithMockUser;
 import org.springframework.test.context.ContextConfiguration;
+import org.springframework.test.context.junit.jupiter.SpringExtension;
 import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.web.context.WebApplicationContext;
@@ -39,7 +42,8 @@ import static org.springframework.test.web.servlet.setup.MockMvcBuilders.webAppC
  * @author Mylene Simon <mylene.simon at nist.gov>
  *
  */
-@RunWith(SpringRunner.class)
+@ExtendWith(SpringExtension.class)
+@TestInstance(TestInstance.Lifecycle.PER_CLASS)
 @ContextConfiguration
 @SpringBootTest(classes = { Application.class, SecurityConfig.class },
         properties = { "spring.data.mongodb.port=0" })
@@ -57,7 +61,7 @@ public class PluginRepositoryRestTest {
 
     Plugin plugin;
 
-    @Before
+    @BeforeAll
     public void setUp() {
         mvc = webAppContextSetup(context)
                 .apply(springSecurity())
