@@ -13,13 +13,14 @@ package gov.nist.itl.ssd.wipp.backend.data.csvcollection;
 
 import static org.hamcrest.Matchers.*;
 import static org.hamcrest.MatcherAssert.assertThat;
-import static org.junit.Assert.*;
 import static org.springframework.security.test.web.servlet.setup.SecurityMockMvcConfigurers.*;
 import static org.springframework.test.web.servlet.setup.MockMvcBuilders.*;
 
-import org.junit.Before;
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.TestInstance;
+import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.data.domain.Page;
@@ -27,8 +28,9 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.security.access.AccessDeniedException;
 import org.springframework.security.test.context.support.WithAnonymousUser;
+import org.springframework.security.test.context.support.WithMockUser;
 import org.springframework.security.web.FilterChainProxy;
-import org.springframework.test.context.junit4.SpringRunner;
+import org.springframework.test.context.junit.jupiter.SpringExtension;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.web.context.WebApplicationContext;
 
@@ -36,7 +38,7 @@ import gov.nist.itl.ssd.wipp.backend.Application;
 import gov.nist.itl.ssd.wipp.backend.app.SecurityConfig;
 import gov.nist.itl.ssd.wipp.backend.data.csvCollection.CsvCollection;
 import gov.nist.itl.ssd.wipp.backend.data.csvCollection.CsvCollectionRepository;
-import gov.nist.itl.ssd.wipp.backend.securityutils.WithMockKeycloakUser;
+//import gov.nist.itl.ssd.wipp.backend.securityutils.WithMockKeycloakUser;
 
 /**
  * Collection of tests for {@link CsvCollectionRepository} exposed methods
@@ -46,7 +48,8 @@ import gov.nist.itl.ssd.wipp.backend.securityutils.WithMockKeycloakUser;
  * @author Mylene Simon <mylene.simon at nist.gov>
  *
  */
-@RunWith(SpringRunner.class)
+@ExtendWith(SpringExtension.class)
+@TestInstance(TestInstance.Lifecycle.PER_CLASS)
 @SpringBootTest(classes = { Application.class, SecurityConfig.class }, 
 				properties = { "spring.data.mongodb.port=0" })
 public class CsvCollectionRepositoryTest {
@@ -61,7 +64,7 @@ public class CsvCollectionRepositoryTest {
 	
 	CsvCollection publicCollA, publicCollB, privateCollA, privateCollB;
 	
-	@Before
+	@BeforeAll
 	public void setUp() {
 		this.mvc = webAppContextSetup(context)
 				.apply(springSecurity())
@@ -103,14 +106,14 @@ public class CsvCollectionRepositoryTest {
 		// Anonymous user should not be able to read a private collection
 		try {
 			csvCollectionRepository.findById(privateCollA.getId());
-			fail("Expected AccessDenied security error");
+			Assertions.fail("Expected AccessDenied security error");
 		} catch (AccessDeniedException e) {
 			// expected
 		}
 	}
 	
 	@Test
-	@WithMockKeycloakUser(username="user1", roles={ "user" })
+	@WithMockUser(username="user1", roles={ "user" })
 	public void findById_nonAdminCallingShouldReturnOnlyOwnOrPublicItems() throws Exception {
 		
 		// Non-admin user1 should be able to read own private collection
@@ -122,14 +125,14 @@ public class CsvCollectionRepositoryTest {
 		// Non-admin user1 should not be able to read a private collection from user2
 		try {
 			csvCollectionRepository.findById(privateCollB.getId());
-			fail("Expected AccessDenied security error");
+			Assertions.fail("Expected AccessDenied security error");
 		} catch (AccessDeniedException e) {
 			// expected
 		}
 	}
 
 	@Test
-	@WithMockKeycloakUser(username="admin", roles={ "admin" })
+	@WithMockUser(username="admin", roles={ "admin" })
 	public void findById_adminCallingShouldReturnAllItems() throws Exception {
 		
 		// Admin should be able to read a public collection from user1
@@ -154,7 +157,7 @@ public class CsvCollectionRepositoryTest {
 	}
 	
 	@Test
-	@WithMockKeycloakUser(username="user1", roles={ "user" })
+	@WithMockUser(username="user1", roles={ "user" })
 	public void findAll_nonAdminCallingShouldReturnOnlyOwnOrPublicItems() throws Exception {
 		
 		Pageable pageable = PageRequest.of(0, 10);
@@ -168,7 +171,7 @@ public class CsvCollectionRepositoryTest {
 	}
 
 	@Test
-	@WithMockKeycloakUser(username="admin", roles={ "admin" })
+	@WithMockUser(username="admin", roles={ "admin" })
 	public void findAll_adminCallingShouldReturnAllItems() throws Exception {
 		
 		Pageable pageable = PageRequest.of(0, 10);
@@ -193,7 +196,7 @@ public class CsvCollectionRepositoryTest {
 	}
 	
 	@Test
-	@WithMockKeycloakUser(username="user1", roles={ "user" })
+	@WithMockUser(username="user1", roles={ "user" })
 	public void findByNameContainingIgnoreCase_nonAdminCallingShouldReturnOnlyOwnOrPublicItems() throws Exception {
 		
 		Pageable pageable = PageRequest.of(0, 10);
@@ -207,7 +210,7 @@ public class CsvCollectionRepositoryTest {
 	}
 
 	@Test
-	@WithMockKeycloakUser(username="admin", roles={ "admin" })
+	@WithMockUser(username="admin", roles={ "admin" })
 	public void findByNameContainingIgnoreCase_adminCallingShouldReturnAllItems() throws Exception {
 		
 		Pageable pageable = PageRequest.of(0, 10);
@@ -239,7 +242,7 @@ public class CsvCollectionRepositoryTest {
 	}
 	
 	@Test
-	@WithMockKeycloakUser(username="user1", roles={ "user" })
+	@WithMockUser(username="user1", roles={ "user" })
 	public void findByName_nonAdminCallingShouldReturnOnlyOwnOrPublicItems() 
 			throws Exception {
 		
@@ -258,7 +261,7 @@ public class CsvCollectionRepositoryTest {
 	}
 
 	@Test
-	@WithMockKeycloakUser(username="admin", roles={ "admin" })
+	@WithMockUser(username="admin", roles={ "admin" })
 	public void findByName_adminCallingShouldReturnAllItems() 
 			throws Exception {
 		
