@@ -14,13 +14,14 @@ package gov.nist.itl.ssd.wipp.backend.data.genericdatacollection;
 import static org.hamcrest.Matchers.hasSize;
 import static org.hamcrest.Matchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
-import static org.junit.Assert.fail;
 import static org.springframework.security.test.web.servlet.setup.SecurityMockMvcConfigurers.springSecurity;
 import static org.springframework.test.web.servlet.setup.MockMvcBuilders.webAppContextSetup;
 
-import org.junit.Before;
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.TestInstance;
+import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.data.domain.Page;
@@ -28,8 +29,9 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.security.access.AccessDeniedException;
 import org.springframework.security.test.context.support.WithAnonymousUser;
+import org.springframework.security.test.context.support.WithMockUser;
 import org.springframework.security.web.FilterChainProxy;
-import org.springframework.test.context.junit4.SpringRunner;
+import org.springframework.test.context.junit.jupiter.SpringExtension;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.web.context.WebApplicationContext;
 
@@ -37,7 +39,7 @@ import gov.nist.itl.ssd.wipp.backend.Application;
 import gov.nist.itl.ssd.wipp.backend.app.SecurityConfig;
 import gov.nist.itl.ssd.wipp.backend.data.genericdatacollection.GenericDataCollection;
 import gov.nist.itl.ssd.wipp.backend.data.genericdatacollection.GenericDataCollectionRepository;
-import gov.nist.itl.ssd.wipp.backend.securityutils.WithMockKeycloakUser;
+//import gov.nist.itl.ssd.wipp.backend.securityutils.WithMockKeycloakUser;
 
 /**
  * Collection of tests for {@link GenericDataCollectionRepository} exposed methods
@@ -47,7 +49,8 @@ import gov.nist.itl.ssd.wipp.backend.securityutils.WithMockKeycloakUser;
  * @author Mylene Simon <mylene.simon at nist.gov>
  *
  */
-@RunWith(SpringRunner.class)
+@ExtendWith(SpringExtension.class)
+@TestInstance(TestInstance.Lifecycle.PER_CLASS)
 @SpringBootTest(classes = { Application.class, SecurityConfig.class }, 
 				properties = { "spring.data.mongodb.port=0" })
 public class GenericDataCollectionRepositoryTest {
@@ -62,7 +65,7 @@ public class GenericDataCollectionRepositoryTest {
 	
 	GenericDataCollection publicGenDataA, publicGenDataB, privateGenDataA, privateGenDataB;
 	
-	@Before
+	@BeforeAll
 	public void setUp() {
 		this.mvc = webAppContextSetup(context)
 				.apply(springSecurity())
@@ -104,14 +107,14 @@ public class GenericDataCollectionRepositoryTest {
 		// Anonymous user should not be able to read a private collection
 		try {
 			genericDataCollectionRepository.findById(privateGenDataA.getId());
-			fail("Expected AccessDenied security error");
+			Assertions.fail("Expected AccessDenied security error");
 		} catch (AccessDeniedException e) {
 			// expected
 		}
 	}
 	
 	@Test
-	@WithMockKeycloakUser(username="user1", roles={ "user" })
+	@WithMockUser(username="user1", roles={ "user" })
 	public void findById_nonAdminCallingShouldReturnOnlyOwnOrPublicItems() throws Exception {
 		
 		// Non-admin user1 should be able to read own private collection
@@ -123,14 +126,14 @@ public class GenericDataCollectionRepositoryTest {
 		// Non-admin user1 should not be able to read a private collection from user2
 		try {
 			genericDataCollectionRepository.findById(privateGenDataB.getId());
-			fail("Expected AccessDenied security error");
+			Assertions.fail("Expected AccessDenied security error");
 		} catch (AccessDeniedException e) {
 			// expected
 		}
 	}
 
 	@Test
-	@WithMockKeycloakUser(username="admin", roles={ "admin" })
+	@WithMockUser(username="admin", roles={ "admin" })
 	public void findById_adminCallingShouldReturnAllItems() throws Exception {
 		
 		// Admin should be able to read a public collection from user1
@@ -155,7 +158,7 @@ public class GenericDataCollectionRepositoryTest {
 	}
 	
 	@Test
-	@WithMockKeycloakUser(username="user1", roles={ "user" })
+	@WithMockUser(username="user1", roles={ "user" })
 	public void findAll_nonAdminCallingShouldReturnOnlyOwnOrPublicItems() throws Exception {
 		
 		Pageable pageable = PageRequest.of(0, 10);
@@ -169,7 +172,7 @@ public class GenericDataCollectionRepositoryTest {
 	}
 
 	@Test
-	@WithMockKeycloakUser(username="admin", roles={ "admin" })
+	@WithMockUser(username="admin", roles={ "admin" })
 	public void findAll_adminCallingShouldReturnAllItems() throws Exception {
 		
 		Pageable pageable = PageRequest.of(0, 10);
@@ -194,7 +197,7 @@ public class GenericDataCollectionRepositoryTest {
 	}
 	
 	@Test
-	@WithMockKeycloakUser(username="user1", roles={ "user" })
+	@WithMockUser(username="user1", roles={ "user" })
 	public void findByNameContainingIgnoreCase_nonAdminCallingShouldReturnOnlyOwnOrPublicItems() throws Exception {
 		
 		Pageable pageable = PageRequest.of(0, 10);
@@ -208,7 +211,7 @@ public class GenericDataCollectionRepositoryTest {
 	}
 
 	@Test
-	@WithMockKeycloakUser(username="admin", roles={ "admin" })
+	@WithMockUser(username="admin", roles={ "admin" })
 	public void findByNameContainingIgnoreCase_adminCallingShouldReturnAllItems() throws Exception {
 		
 		Pageable pageable = PageRequest.of(0, 10);
