@@ -22,7 +22,8 @@ import java.util.Optional;
 import java.util.zip.ZipEntry;
 import java.util.zip.ZipOutputStream;
 
-import javax.servlet.http.HttpServletResponse;
+import io.swagger.v3.oas.annotations.tags.Tag;
+import jakarta.servlet.http.HttpServletResponse;
 
 import org.apache.commons.io.IOUtils;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -37,14 +38,13 @@ import gov.nist.itl.ssd.wipp.backend.core.model.data.DataDownloadTokenRepository
 import gov.nist.itl.ssd.wipp.backend.core.rest.DownloadUrl;
 import gov.nist.itl.ssd.wipp.backend.core.rest.exception.ForbiddenException;
 import gov.nist.itl.ssd.wipp.backend.core.utils.SecurityUtils;
-import io.swagger.annotations.Api;
 
 /**
 *
 * @author Mohamed Ouladi <mohamed.ouladi at labshare.org>
 */
 @RestController
-@Api(tags="GenericDataCollection Entity")
+@Tag(name="GenericDataCollection Entity")
 @RequestMapping(CoreConfig.BASE_URI + "/genericDataCollections/{genericDataCollectionId}/download")
 public class GenericDataCollectionDownloadController {
 
