@@ -25,8 +25,8 @@ import gov.nist.itl.ssd.wipp.backend.core.model.workflow.WorkflowRepository;
 import gov.nist.itl.ssd.wipp.backend.core.model.workflow.WorkflowStatus;
 import gov.nist.itl.ssd.wipp.backend.core.rest.exception.ClientException;
 import gov.nist.itl.ssd.wipp.backend.core.utils.SecurityUtils;
-import io.swagger.annotations.Api;
 
+import io.swagger.v3.oas.annotations.tags.Tag;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -45,7 +45,7 @@ import java.util.Optional;
  * @author Mylene Simon <mylene.simon at nist.gov>
  */
 @Controller
-@Api(tags="Workflow Entity")
+@Tag(name="Workflow Entity")
 @RequestMapping(CoreConfig.BASE_URI + "/workflows/{workflowId}/exit")
 public class WorkflowExitController {
 
