@@ -25,7 +25,7 @@ import java.util.logging.Level;
 import java.util.logging.Logger;
 
 import org.apache.commons.io.FileUtils;
-import org.apache.commons.lang.StringUtils;
+import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.rest.core.annotation.*;
 import org.springframework.security.access.prepost.PreAuthorize;
@@ -73,6 +73,7 @@ public class ImagesCollectionEventHandler {
 
         // Set the owner to the connected user
         imagesCollection.setOwner(SecurityContextHolder.getContext().getAuthentication().getName());
+        System.out.println(SecurityContextHolder.getContext().getAuthentication().getAuthorities());
 
         
         // Default import method is UPLOADED
@@ -99,7 +100,7 @@ public class ImagesCollectionEventHandler {
         }
     }
     @HandleBeforeSave
-    @PreAuthorize("isAuthenticated() and (hasRole('admin') or #imagesCollection.owner == principal.name)")
+    @PreAuthorize("isAuthenticated() and (hasRole('admin') or #imagesCollection.owner == authentication.name)")
     public void handleBeforeSave(ImagesCollection imagesCollection) {
     	// Assert collection exists
         Optional<ImagesCollection> result = imagesCollectionRepository.findById(
@@ -173,7 +174,7 @@ public class ImagesCollectionEventHandler {
 
     @HandleBeforeDelete
     @PreAuthorize("isAuthenticated() and (hasRole('admin') or "
-    		+ "(#imagesCollection.owner == principal.name and #imagesCollection.publiclyShared == false))")
+    		+ "(#imagesCollection.owner == authentication.name and #imagesCollection.publiclyShared == false))")
     public void handleBeforeDelete(ImagesCollection imagesCollection) {
     	// Assert collection exists
     	Optional<ImagesCollection> result = imagesCollectionRepository.findById(
