@@ -12,7 +12,7 @@
 package gov.nist.itl.ssd.wipp.backend.data.imagescollection;
 
 import gov.nist.itl.ssd.wipp.backend.core.CoreConfig;
-import org.apache.commons.lang.StringUtils;
+import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
