@@ -14,6 +14,7 @@ package gov.nist.itl.ssd.wipp.backend.data.jupyternotebook;
 import java.io.File;
 import java.io.IOException;
 
+import io.swagger.v3.oas.annotations.tags.Tag;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -23,14 +24,13 @@ import org.springframework.web.bind.annotation.RestController;
 
 import gov.nist.itl.ssd.wipp.backend.core.CoreConfig;
 import gov.nist.itl.ssd.wipp.backend.core.rest.exception.ClientException;
-import io.swagger.annotations.Api;
 
 /**
 *
 * @author Mohamed Ouladi <mohamed.ouladi at nist.gov>
 */
 @RestController
-@Api(tags="Notebook Entity")
+@Tag(name="Notebook Entity")
 @RequestMapping(CoreConfig.BASE_URI + "/notebooks/import")
 public class NotebookImportController {
 
